requirements.txt
discord.py
yt-dlp
google-api-python-client
python-dotenv
PyNaCl>=1.5.0
lyricsgenius
flask